error en swagger rectificado

+++ src/main/java/com/business/client/service/model/dto/AddClientDTO.java
@@ -17,7 +17,7 @@ public class AddClientDTO {
     private String phone;
     private String address;
     private Integer typeClient;
-    private Byte result;
+    private Integer result;
     private String errorMessage;
